% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AF_PER_CFS_DAY}
\alias{AF_PER_CFS_DAY}
\alias{CFS_DAY_PER_AF}
\alias{CFS_PER_MGD}
\alias{MGD_PER_CFS}
\title{Hydrologic constants}
\format{An object of class \code{numeric} of length 1.}
\usage{
AF_PER_CFS_DAY
}
\description{
used for converting between flows in (k)cfs and volumes in (K)AF or between cfs and MGD
}
\note{
acre-feet to CFS-day is an exact conversion as cubic feet per acre-foot is 43560, seconds in a day is 86400, simplifing the conversion gets 24/12.1
}
\author{
Evan Heisman
}
\keyword{datasets}

