% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroplots.R
\name{hydro_axis_trans}
\alias{flowBreaks}
\alias{hydro_axis_trans}
\alias{hydro_flow_breaks}
\alias{hydro_flow_trans}
\alias{hydro_prob_breaks}
\alias{hydro_prob_trans}
\alias{probBreaks}
\title{Hydrologic axis transforms for ggplot2}
\usage{
flowBreaks(Q, labels = c(1, 2, 3, 5, 7))
}
\arguments{
\item{Q}{values for flow to create scale along}

\item{labels}{First digit of breaks to label.  Defaults to 1,2,3,5,7 for log axes; 1,2,5 for probability axes.  Multiples of these are assumed.}

\item{maxLevel}{smallest place to which breaks should be created for probability axes.}

\item{lines}{which breaks should have lines for probability axes, must be superset of labels}

\item{invert}{switch the direction of the probability axes}

\item{as.percent}{Show labels by percent instead of probability}

\item{byPeriod}{Show labels with return interval (1 / probability)}

\item{periodSuffix}{String to concatenate to return interval}

\item{distribution}{change distribution for probability axis, defaults to normal.  Untested with other distributions.}

\item{distArgs}{arguments pasted to distribution specified in \code{distribution}}
}
\description{
Axis transforms for hydrologic plots in ggplot2.
}
\details{
\code{hydro_prob_trans} and \code{hydro_prob_breaks} for probability axes on frequency plots

\code{hydro_flow_trans} and \code{hydro_flow_breaks} for log axes with additional breaks between powers of ten.
}
\examples{
ggplot(peaks, aes(y=PEAK_DAILY_FLOW, x=weibullProbs(PEAK_DAILY_FLOW))) + geom_point() +
  theme_bw(base_size=11) + theme(legend.position = "bottom", panel.grid.minor=element_blank()) +
  scale_y_continuous(trans=hydro_flow_trans()) +
  scale_x_continuous(trans=hydro_prob_trans(lines=c(1,2,5), labels=c(1,2,5), byPeriod=TRUE)) +
  stat_smooth(method="glm", family=gaussian(link="log"))
}
\author{
Evan Heisman
}

