% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroplots.R
\name{mergePlots}
\alias{mergePlots}
\title{Merge two ggplot plots with a legend on the right hand side such that the plots have the same width.}
\usage{
mergePlots(..., plotTitle = "", heights = NULL)
}
\arguments{
\item{...}{ggplot grobs to merge}

\item{plotTitle}{title to place at top of plot}

\item{heights}{vector of weights (e.g. \code{c(2,1,1)} for half page and two quarter page plots) for heights}
}
\value{
grob of merged plots
}
\description{
Used for merging plots of different heights (unlike facets) with a common x-axis.
}
\author{
Evan Heisman
}

