% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{plottingPosition}
\alias{plottingPosition}
\alias{weibullProbs}
\title{Compute plotting positions for flow frequency curves}
\usage{
\code{plottingPosition(flows)}
\code{weibullPlottingPosition(flows)}
}
\arguments{
\item{Qs}{list of points for which weibull plotting positions are needed}

\item{a}{Plotting position parameter.  0 for Weibull (default), 0.5 for Hazen,}

\item{exceedance}{boolean that determines if plotting positions should be exceedance probabilities or non-exceedance probabilities.}

\item{as.points}{boolean if ties should get independent points or not.}
}
\value{
probabilities in range of (0,1) corresponding to each point in input vector.
}
\description{
generates weibull plotting positions for given vector.  Often used in hydrologic 'flow frequency curves'
}
\note{
\code{weibullProbs} provided for backwards compatibility.
}
\author{
Evan Heisman
}

